Commit for further validation
This push has functionality for lti validation as well as some clean up in  the other stages of validation, currently working on a way to implement iframe validation to searchfor splice messages using pupateer. Currently the lti microservice api I built works but we need a way to get keys and secrets, for now I hardedcoded a known working lti url and key/secret.